+++ app/Providers/AppServiceProvider.php
@@ -2,6 +2,13 @@
 
 namespace App\Providers;
 
+use App\Models\ModuleRequest;
+use App\Models\Role;
+use App\Models\User;
+use App\Policies\ModuleRequestPolicy;
+use App\Policies\RolePolicy;
+use App\Policies\UserPolicy;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\ServiceProvider;
 
 class AppServiceProvider extends ServiceProvider
@@ -19,6 +26,8 @@ class AppServiceProvider extends ServiceProvider
      */
     public function boot(): void
     {
-        //
+        Gate::policy(ModuleRequest::class, ModuleRequestPolicy::class);
+        Gate::policy(User::class, UserPolicy::class);
+        Gate::policy(Role::class, RolePolicy::class);
     }
 }